ebano_app/requirements.txt
Flask==2.2.5
Jinja2==3.1.2
Werkzeug==2.2.3
gunicorn==20.1.0
pg8000==1.29.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Faker==18.3.1
requests==2.31.0
bcrypt==4.0.1
itsdangerous==2.1.2
PyJWT==2.8.0
email-validator==2.1.0
Flask-Limiter==3.5.0